requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.5
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
Pillow==9.4.0
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.3
twilio==8.2.2
urllib3==2.0.3
yarl==1.9.2
